Seperated out all of the Expression and ExpressionBuilder classes into their own files.
