+++ lib/query/Expression.js
@@ -6,9 +6,15 @@ Object.defineProperty(exports, "__esModule", {
 
 var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
 
-function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+var _OperationExpression = require("./OperationExpression");
 
-function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+var _OperationExpression2 = _interopRequireDefault(_OperationExpression);
+
+var _ValueExpression = require("./ValueExpression");
+
+var _ValueExpression2 = _interopRequireDefault(_ValueExpression);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
@@ -58,87 +64,87 @@ var Expression = function () {
     }, {
         key: "property",
         value: function property(value) {
-            return new ValueExpression("property", value);
+            return new _ValueExpression2.default("property", value);
         }
     }, {
         key: "constant",
         value: function constant(value) {
-            return new ValueExpression("constant", value);
+            return new _ValueExpression2.default("constant", value);
         }
     }, {
         key: "boolean",
         value: function boolean(value) {
-            var expression = new ValueExpression("boolean");
+            var expression = new _ValueExpression2.default("boolean");
             expression.value = value;
             return expression;
         }
     }, {
         key: "string",
         value: function string(value) {
-            var expression = new ValueExpression("string");
+            var expression = new _ValueExpression2.default("string");
             expression.value = value;
             return expression;
         }
     }, {
         key: "number",
         value: function number(value) {
-            var expression = new ValueExpression("number");
+            var expression = new _ValueExpression2.default("number");
             expression.value = value;
             return expression;
         }
     }, {
         key: "object",
         value: function object(value) {
-            var expression = new ValueExpression("object");
+            var expression = new _ValueExpression2.default("object");
             expression.value = value;
             return expression;
         }
     }, {
         key: "date",
         value: function date(value) {
-            var expression = new ValueExpression("date");
+            var expression = new _ValueExpression2.default("date");
             expression.value = value;
             return expression;
         }
     }, {
         key: "function",
         value: function _function(value) {
-            var expression = new ValueExpression("function");
+            var expression = new _ValueExpression2.default("function");
             expression.value = value;
             return expression;
         }
     }, {
         key: "type",
         value: function type(value) {
-            var expression = new ValueExpression("type");
+            var expression = new _ValueExpression2.default("type");
             expression.value = value || Object;
             return expression;
         }
     }, {
         key: "null",
         value: function _null(value) {
-            var expression = new ValueExpression("null");
+            var expression = new _ValueExpression2.default("null");
             expression.value = value;
             return expression;
         }
     }, {
         key: "undefined",
         value: function undefined(value) {
-            var expression = new ValueExpression("undefined");
+            var expression = new _ValueExpression2.default("undefined");
             expression.value = value;
             return expression;
         }
     }, {
         key: "array",
         value: function array(value) {
-            var expression = new ValueExpression("array");
+            var expression = new _ValueExpression2.default("array");
             expression.value = value;
             return expression;
         }
     }, {
         key: "queryable",
         value: function queryable(leftExpression, rightExpression) {
-            var expression = new OperationExpression("queryable");
+            var expression = new _OperationExpression2.default("queryable");
             expression.children.push(leftExpression, rightExpression);
             return expression;
         }
@@ -150,7 +156,7 @@ var Expression = function () {
     }, {
         key: "equalTo",
         value: function equalTo() {
-            var expression = new OperationExpression("equalTo");
+            var expression = new _OperationExpression2.default("equalTo");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -159,7 +165,7 @@ var Expression = function () {
     }, {
         key: "notEqualTo",
         value: function notEqualTo() {
-            var expression = new OperationExpression("notEqualTo");
+            var expression = new _OperationExpression2.default("notEqualTo");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -168,7 +174,7 @@ var Expression = function () {
     }, {
         key: "or",
         value: function or() {
-            var expression = new OperationExpression("or");
+            var expression = new _OperationExpression2.default("or");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -177,7 +183,7 @@ var Expression = function () {
     }, {
         key: "and",
         value: function and() {
-            var expression = new OperationExpression("and");
+            var expression = new _OperationExpression2.default("and");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -186,7 +192,7 @@ var Expression = function () {
     }, {
         key: "where",
         value: function where() {
-            var expression = new OperationExpression("where");
+            var expression = new _OperationExpression2.default("where");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -195,7 +201,7 @@ var Expression = function () {
     }, {
         key: "greaterThan",
         value: function greaterThan() {
-            var expression = new OperationExpression("greaterThan");
+            var expression = new _OperationExpression2.default("greaterThan");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -204,7 +210,7 @@ var Expression = function () {
     }, {
         key: "lessThan",
         value: function lessThan() {
-            var expression = new OperationExpression("lessThan");
+            var expression = new _OperationExpression2.default("lessThan");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -213,7 +219,7 @@ var Expression = function () {
     }, {
         key: "greaterThanOrEqualTo",
         value: function greaterThanOrEqualTo() {
-            var expression = new OperationExpression("greaterThanOrEqualTo");
+            var expression = new _OperationExpression2.default("greaterThanOrEqualTo");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -222,7 +228,7 @@ var Expression = function () {
     }, {
         key: "lessThanOrEqualTo",
         value: function lessThanOrEqualTo() {
-            var expression = new OperationExpression("lessThanOrEqualTo");
+            var expression = new _OperationExpression2.default("lessThanOrEqualTo");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -231,7 +237,7 @@ var Expression = function () {
     }, {
         key: "orderBy",
         value: function orderBy() {
-            var expression = new OperationExpression("orderBy");
+            var expression = new _OperationExpression2.default("orderBy");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -240,7 +246,7 @@ var Expression = function () {
     }, {
         key: "descending",
         value: function descending() {
-            var expression = new OperationExpression("descending");
+            var expression = new _OperationExpression2.default("descending");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -249,7 +255,7 @@ var Expression = function () {
     }, {
         key: "ascending",
         value: function ascending() {
-            var expression = new OperationExpression("ascending");
+            var expression = new _OperationExpression2.default("ascending");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -258,7 +264,7 @@ var Expression = function () {
     }, {
         key: "skip",
         value: function skip(value) {
-            var expression = new OperationExpression("skip");
+            var expression = new _OperationExpression2.default("skip");
             var valueExpression = Expression.constant(value);
             expression.children.push(valueExpression);
 
@@ -267,7 +273,7 @@ var Expression = function () {
     }, {
         key: "take",
         value: function take(value) {
-            var expression = new OperationExpression("take");
+            var expression = new _OperationExpression2.default("take");
             var valueExpression = Expression.constant(value);
             expression.children.push(valueExpression);
 
@@ -276,7 +282,7 @@ var Expression = function () {
     }, {
         key: "buildOperatorExpression",
         value: function buildOperatorExpression(name) {
-            var expression = new OperationExpression(name);
+            var expression = new _OperationExpression2.default(name);
             var args = Array.prototype.slice.call(arguments, 1);
             args.forEach(function (arg) {
                 expression.children.push(arg);
@@ -287,7 +293,7 @@ var Expression = function () {
     }, {
         key: "guid",
         value: function guid() {
-            var expression = new OperationExpression("guid");
+            var expression = new _OperationExpression2.default("guid");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -296,7 +302,7 @@ var Expression = function () {
     }, {
         key: "substring",
         value: function substring() {
-            var expression = new OperationExpression("substring");
+            var expression = new _OperationExpression2.default("substring");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -305,7 +311,7 @@ var Expression = function () {
     }, {
         key: "substringOf",
         value: function substringOf() {
-            var expression = new OperationExpression("substringOf");
+            var expression = new _OperationExpression2.default("substringOf");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -314,7 +320,7 @@ var Expression = function () {
     }, {
         key: "startsWith",
         value: function startsWith() {
-            var expression = new OperationExpression("startsWith");
+            var expression = new _OperationExpression2.default("startsWith");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -323,7 +329,7 @@ var Expression = function () {
     }, {
         key: "endsWith",
         value: function endsWith() {
-            var expression = new OperationExpression("endsWith");
+            var expression = new _OperationExpression2.default("endsWith");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -332,7 +338,7 @@ var Expression = function () {
     }, {
         key: "isIn",
         value: function isIn(property, array) {
-            var expression = new OperationExpression("isIn");
+            var expression = new _OperationExpression2.default("isIn");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -341,7 +347,7 @@ var Expression = function () {
     }, {
         key: "isNotIn",
         value: function isNotIn(property, array) {
-            var expression = new OperationExpression("isNotIn");
+            var expression = new _OperationExpression2.default("isNotIn");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -350,7 +356,7 @@ var Expression = function () {
     }, {
         key: "include",
         value: function include() {
-            var expression = new OperationExpression("include");
+            var expression = new _OperationExpression2.default("include");
             Array.prototype.slice.call(arguments, 0).forEach(function (arg) {
                 expression.children.push(arg);
             });
@@ -359,7 +365,7 @@ var Expression = function () {
     }, {
         key: "any",
         value: function any(propertyAccessExpression, expression) {
-            var anyExpression = new OperationExpression("any");
+            var anyExpression = new _OperationExpression2.default("any");
             var expressionExpression = Expression.expression(expression);
 
             anyExpression.children.push(propertyAccessExpression, expressionExpression);
@@ -368,7 +374,7 @@ var Expression = function () {
     }, {
         key: "all",
         value: function all(propertyAccessExpression, expression) {
-            var allExpression = new OperationExpression("all");
+            var allExpression = new _OperationExpression2.default("all");
             var expressionExpression = Expression.expression(expression);
 
             allExpression.children.push(propertyAccessExpression, expressionExpression);
@@ -377,7 +383,7 @@ var Expression = function () {
     }, {
         key: "expression",
         value: function expression(value) {
-            var expresssionExpression = new ValueExpression("expression", value);
+            var expresssionExpression = new _ValueExpression2.default("expression", value);
 
             return expresssionExpression;
         }
@@ -385,7 +391,7 @@ var Expression = function () {
         key: "propertyAccess",
         value: function propertyAccess(leftExpression, propertyName) {
             var propertyExpression = Expression.property(propertyName);
-            var propertyAccessExpression = new OperationExpression("propertyAccess");
+            var propertyAccessExpression = new _OperationExpression2.default("propertyAccess");
             propertyAccessExpression.children.push(leftExpression, propertyExpression);
 
             return propertyAccessExpression;
@@ -393,9 +399,9 @@ var Expression = function () {
     }, {
         key: "contains",
         value: function contains(Type, namespace, expression) {
-            var containsExpression = new OperationExpression("contains");
-            var ofTypeExpression = new ValueExpression("ofType", Type);
-            var propertyExpression = new ValueExpression("property", namespace);
+            var containsExpression = new _OperationExpression2.default("contains");
+            var ofTypeExpression = new _ValueExpression2.default("ofType", Type);
+            var propertyExpression = new _ValueExpression2.default("property", namespace);
 
             containsExpression.children.push(ofTypeExpression, propertyExpression, expression);
 
@@ -404,9 +410,9 @@ var Expression = function () {
     }, {
         key: "intersects",
         value: function intersects(Type, namespace, expression) {
-            var intersectsExpression = new OperationExpression("intersects");
-            var ofTypeExpression = new ValueExpression("ofType", Type);
-            var propertyExpression = new ValueExpression("property", namespace);
+            var intersectsExpression = new _OperationExpression2.default("intersects");
+            var ofTypeExpression = new _ValueExpression2.default("ofType", Type);
+            var propertyExpression = new _ValueExpression2.default("property", namespace);
 
             intersectsExpression.children.push(ofTypeExpression, propertyExpression, expression);
 
@@ -417,134 +423,5 @@ var Expression = function () {
     return Expression;
 }();
 
-var ValueExpression = function (_Expression) {
-    _inherits(ValueExpression, _Expression);
-
-    function ValueExpression(nodeName, value) {
-        _classCallCheck(this, ValueExpression);
-
-        var _this = _possibleConstructorReturn(this, (ValueExpression.__proto__ || Object.getPrototypeOf(ValueExpression)).call(this));
-
-        _this.value = value;
-        _this.nodeName = nodeName;
-        return _this;
-    }
-
-    _createClass(ValueExpression, [{
-        key: "copy",
-        value: function copy() {
-            return new ValueExpression(this.nodeName, this.value);
-        }
-    }, {
-        key: "isEqualTo",
-        value: function isEqualTo(node) {
-            if (node && this.nodeName === node.nodeName && this.value === node.value) {
-                return true;
-            }
-            return false;
-        }
-    }, {
-        key: "contains",
-        value: function contains(node) {
-            return this.isEqualTo(node);
-        }
-    }]);
-
-    return ValueExpression;
-}(Expression);
-
-var OperationExpression = function (_Expression2) {
-    _inherits(OperationExpression, _Expression2);
-
-    function OperationExpression(nodeName) {
-        _classCallCheck(this, OperationExpression);
-
-        var _this2 = _possibleConstructorReturn(this, (OperationExpression.__proto__ || Object.getPrototypeOf(OperationExpression)).call(this));
-
-        var args = Array.prototype.slice.call(arguments, 0);
-
-        _this2.nodeName = nodeName;
-        _this2.children = args.slice(1);
-        return _this2;
-    }
-
-    _createClass(OperationExpression, [{
-        key: "copy",
-        value: function copy() {
-            var children = [];
-            var copy = new OperationExpression(this.nodeName);
-
-            this.children.forEach(function (expression) {
-                copy.children.push(expression.copy());
-            });
-
-            return copy;
-        }
-    }, {
-        key: "isEqualTo",
-        value: function isEqualTo() {
-            if (!Array.isArray(node.children) || this.nodeName !== node.nodeName) {
-                return false;
-            }
-
-            if (node.children.length !== this.children.length) {
-                return false;
-            }
-
-            return this.children.every(function (expression, index) {
-                return expression.isEqualTo(node.children[index]);
-            });
-        }
-    }, {
-        key: "contains",
-        value: function contains(node) {
-            var _this3 = this;
-
-            if (node.nodeName === this.nodeName && Array.isArray(node.children)) {
-                var matched = node.children.every(function (childNode, index) {
-                    return childNode.contains(_this3.children[index]);
-                });
-
-                if (matched) {
-                    return true;
-                }
-            }
-
-            return this.children.some(function (childNode) {
-                return childNode.contains(node);
-            });
-        }
-    }, {
-        key: "getMatchingNodes",
-        value: function getMatchingNodes(node, matchedNodes) {
-            var _this4 = this;
-
-            matchedNodes = Array.isArray(matchedNodes) ? matchedNodes : [];
-
-            if (node.nodeName === this.nodeName && Array.isArray(node.children)) {
-                var matched = node.children.every(function (childNode, index) {
-                    return childNode.contains(_this4.children[index], matchedNodes);
-                });
-
-                if (matched) {
-                    matchedNodes.push(this);
-                }
-            }
-
-            this.children.forEach(function (childNode) {
-                if (Array.isArray(childNode.children)) {
-                    childNode.getMatchingNodes(node, matchedNodes);
-                }
-            }, matchedNodes);
-
-            return matchedNodes;
-        }
-    }]);
-
-    return OperationExpression;
-}(Expression);
-
-exports.Expression = Expression;
-exports.ValueExpression = ValueExpression;
-exports.OperationExpression = OperationExpression;
+exports.default = Expression;
 //# sourceMappingURL=Expression.js.map
