+++ lib/query/OperationExpression.js
@@ -1,0 +1,113 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+    value: true
+});
+
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
+
+var _Expression2 = require("./Expression");
+
+var _Expression3 = _interopRequireDefault(_Expression2);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var OperationExpression = function (_Expression) {
+    _inherits(OperationExpression, _Expression);
+
+    function OperationExpression(nodeName) {
+        _classCallCheck(this, OperationExpression);
+
+        var _this = _possibleConstructorReturn(this, (OperationExpression.__proto__ || Object.getPrototypeOf(OperationExpression)).call(this));
+
+        var args = Array.prototype.slice.call(arguments, 0);
+
+        _this.nodeName = nodeName;
+        _this.children = args.slice(1);
+        return _this;
+    }
+
+    _createClass(OperationExpression, [{
+        key: "copy",
+        value: function copy() {
+            var children = [];
+            var copy = new OperationExpression(this.nodeName);
+
+            this.children.forEach(function (expression) {
+                copy.children.push(expression.copy());
+            });
+
+            return copy;
+        }
+    }, {
+        key: "isEqualTo",
+        value: function isEqualTo() {
+            if (!Array.isArray(node.children) || this.nodeName !== node.nodeName) {
+                return false;
+            }
+
+            if (node.children.length !== this.children.length) {
+                return false;
+            }
+
+            return this.children.every(function (expression, index) {
+                return expression.isEqualTo(node.children[index]);
+            });
+        }
+    }, {
+        key: "contains",
+        value: function contains(node) {
+            var _this2 = this;
+
+            if (node.nodeName === this.nodeName && Array.isArray(node.children)) {
+                var matched = node.children.every(function (childNode, index) {
+                    return childNode.contains(_this2.children[index]);
+                });
+
+                if (matched) {
+                    return true;
+                }
+            }
+
+            return this.children.some(function (childNode) {
+                return childNode.contains(node);
+            });
+        }
+    }, {
+        key: "getMatchingNodes",
+        value: function getMatchingNodes(node, matchedNodes) {
+            var _this3 = this;
+
+            matchedNodes = Array.isArray(matchedNodes) ? matchedNodes : [];
+
+            if (node.nodeName === this.nodeName && Array.isArray(node.children)) {
+                var matched = node.children.every(function (childNode, index) {
+                    return childNode.contains(_this3.children[index], matchedNodes);
+                });
+
+                if (matched) {
+                    matchedNodes.push(this);
+                }
+            }
+
+            this.children.forEach(function (childNode) {
+                if (Array.isArray(childNode.children)) {
+                    childNode.getMatchingNodes(node, matchedNodes);
+                }
+            }, matchedNodes);
+
+            return matchedNodes;
+        }
+    }]);
+
+    return OperationExpression;
+}(_Expression3.default);
+
+exports.default = OperationExpression;
+//# sourceMappingURL=OperationExpression.js.map
