+++ lib/query/Queryable.js
@@ -8,7 +8,25 @@ var _createClass = function () { function defineProperties(target, props) { for 
 
 var _Expression = require("./Expression");
 
+var _Expression2 = _interopRequireDefault(_Expression);
+
+var _ValueExpression = require("./ValueExpression");
+
+var _ValueExpression2 = _interopRequireDefault(_ValueExpression);
+
+var _OperationExpression = require("./OperationExpression");
+
+var _OperationExpression2 = _interopRequireDefault(_OperationExpression);
+
 var _ExpressionBuilder = require("./ExpressionBuilder");
+
+var _ExpressionBuilder2 = _interopRequireDefault(_ExpressionBuilder);
+
+var _OperationExpressionBuilder = require("./OperationExpressionBuilder");
+
+var _OperationExpressionBuilder2 = _interopRequireDefault(_OperationExpressionBuilder);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
@@ -45,31 +63,31 @@ var Queryable = function () {
         if (query.where != null && query.where.nodeName === "where") {
             this.query.where = query.where;
         } else {
-            this.query.where = _Expression.Expression.where();
+            this.query.where = _Expression2.default.where();
         }
 
         if (query.skip != null && query.skip.nodeName === "skip") {
             this.query.skip = query.skip;
         } else {
-            this.query.skip = _Expression.Expression.skip(0);
+            this.query.skip = _Expression2.default.skip(0);
         }
 
         if (query.take != null && query.take.nodeName === "take") {
             this.query.take = query.take;
         } else {
-            this.query.take = _Expression.Expression.take(Infinity);
+            this.query.take = _Expression2.default.take(Infinity);
         }
 
         if (query.include != null && query.include.nodeName === "include") {
             this.query.include = query.include;
         } else {
-            this.query.include = _Expression.Expression.include();
+            this.query.include = _Expression2.default.include();
         }
 
         if (query.orderBy != null && query.orderBy.nodeName === "orderBy") {
             this.query.orderBy = query.orderBy;
         } else {
-            this.query.orderBy = _Expression.Expression.orderBy();
+            this.query.orderBy = _Expression2.default.orderBy();
         }
     }
 
@@ -91,8 +109,8 @@ var Queryable = function () {
 
             if (typeof lambda === "function") {
                 lambda = lambda || function () {};
-                rightExpression = lambda.call(_ExpressionBuilder.ExpressionBuilder, new _ExpressionBuilder.ExpressionBuilder(this.Type));
-            } else if (lambda instanceof _Expression.Expression) {
+                rightExpression = lambda.call(_ExpressionBuilder2.default, new _ExpressionBuilder2.default(this.Type));
+            } else if (lambda instanceof _Expression2.default) {
                 rightExpression = lambda;
             } else {
                 return;
@@ -102,7 +120,7 @@ var Queryable = function () {
                 query.where.children.push(rightExpression);
             } else {
                 var leftExpression = query.where.children.pop();
-                query.where.children.push(_Expression.Expression.or(leftExpression, rightExpression));
+                query.where.children.push(_Expression2.default.or(leftExpression, rightExpression));
             }
 
             return this.copy(query);
@@ -115,8 +133,8 @@ var Queryable = function () {
 
             if (typeof lambda === "function") {
                 lambda = lambda || function () {};
-                rightExpression = lambda.call(_ExpressionBuilder.ExpressionBuilder, new _ExpressionBuilder.ExpressionBuilder(this.Type));
-            } else if (lambda instanceof _Expression.Expression) {
+                rightExpression = lambda.call(_ExpressionBuilder2.default, new _ExpressionBuilder2.default(this.Type));
+            } else if (lambda instanceof _Expression2.default) {
                 rightExpression = lambda;
             } else {
                 throw new Error("Expected an expression to be supplied.");
@@ -126,7 +144,7 @@ var Queryable = function () {
                 query.where.children.push(rightExpression);
             } else {
                 var leftExpression = query.where.children.pop();
-                query.where.children.push(_Expression.Expression.and(leftExpression, rightExpression));
+                query.where.children.push(_Expression2.default.and(leftExpression, rightExpression));
             }
 
             return this.copy(query);
@@ -144,7 +162,7 @@ var Queryable = function () {
             }
 
             var query = copyQuery(this.getQuery());
-            query.take = _Expression.Expression.take(value);
+            query.take = _Expression2.default.take(value);
 
             return this.copy(query);
         }
@@ -156,7 +174,7 @@ var Queryable = function () {
             }
 
             var query = copyQuery(this.getQuery());
-            query.skip = _Expression.Expression.skip(value);
+            query.skip = _Expression2.default.skip(value);
 
             return this.copy(query);
         }
@@ -164,9 +182,9 @@ var Queryable = function () {
         key: "orderByDesc",
         value: function orderByDesc(lambda) {
             var query = copyQuery(this.getQuery());
-            var propertyExpression = lambda.call(_ExpressionBuilder.ExpressionBuilder, new _ExpressionBuilder.ExpressionBuilder(this.Type)).getExpression();
+            var propertyExpression = lambda.call(_ExpressionBuilder2.default, new _ExpressionBuilder2.default(this.Type)).getExpression();
 
-            var descendingExpression = _Expression.Expression.descending(propertyExpression);
+            var descendingExpression = _Expression2.default.descending(propertyExpression);
 
             if (!query.orderBy.contains(propertyExpression)) {
                 query.orderBy.children.push(descendingExpression);
@@ -183,14 +201,14 @@ var Queryable = function () {
 
             if (typeof lambda === "function") {
                 lambda = lambda || function () {};
-                propertyExpression = lambda.call(_ExpressionBuilder.ExpressionBuilder, new _ExpressionBuilder.ExpressionBuilder(this.Type)).getExpression();
-            } else if (lambda instanceof _ExpressionBuilder.OperationExpressionBuilder) {
+                propertyExpression = lambda.call(_ExpressionBuilder2.default, new _ExpressionBuilder2.default(this.Type)).getExpression();
+            } else if (lambda instanceof _OperationExpressionBuilder2.default) {
                 propertyExpression = lambda.getExpression();
             } else {
                 throw new Error("Expected a property to orderBy.");
             }
 
-            var ascendingExpression = _Expression.Expression.ascending(propertyExpression);
+            var ascendingExpression = _Expression2.default.ascending(propertyExpression);
 
             if (!query.orderBy.contains(propertyExpression)) {
                 query.orderBy.children.push(ascendingExpression);
@@ -233,15 +251,15 @@ var Queryable = function () {
 
             if (typeof lambda === "function") {
                 lambda = lambda || function () {};
-                propertyExpression = lambda.call(_ExpressionBuilder.ExpressionBuilder, new _ExpressionBuilder.ExpressionBuilder(this.Type)).getExpression();
-            } else if (lambda instanceof _ExpressionBuilder.OperationExpressionBuilder) {
+                propertyExpression = lambda.call(_ExpressionBuilder2.default, new _ExpressionBuilder2.default(this.Type)).getExpression();
+            } else if (lambda instanceof _OperationExpressionBuilder2.default) {
                 propertyExpression = lambda.getExpression();
             } else {
                 throw new Error("Expected a property to include.");
             }
 
             if (propertyExpression.nodeName !== "queryable") {
-                propertyExpression = _Expression.Expression.queryable(propertyExpression, _Expression.Expression.expression(_Expression.Expression.where()));
+                propertyExpression = _Expression2.default.queryable(propertyExpression, _Expression2.default.expression(_Expression2.default.where()));
             }
 
             query.include.children.push(propertyExpression);
@@ -263,7 +281,7 @@ var Queryable = function () {
                     cloneQuery.where.children[0].children.push(rightExpression.copy());
                 } else {
                     var leftExpression = cloneQuery.where.children.pop();
-                    cloneQuery.where.children.push(_Expression.Expression.and(leftExpression, rightExpression.copy()));
+                    cloneQuery.where.children.push(_Expression2.default.and(leftExpression, rightExpression.copy()));
                 }
             }
 
@@ -288,7 +306,7 @@ var Queryable = function () {
     }, {
         key: "toGuid",
         value: function toGuid(value) {
-            return _Expression.Expression.guid(_Expression.Expression.constant(value));
+            return _Expression2.default.guid(_Expression2.default.constant(value));
         }
     }, {
         key: "forEach",
