+++ lib/query/ValueExpression.js
@@ -1,0 +1,58 @@
+"use strict";
+
+Object.defineProperty(exports, "__esModule", {
+    value: true
+});
+
+var _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();
+
+var _Expression2 = require("./Expression");
+
+var _Expression3 = _interopRequireDefault(_Expression2);
+
+function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
+
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return call && (typeof call === "object" || typeof call === "function") ? call : self; }
+
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function, not " + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
+
+var ValueExpression = function (_Expression) {
+    _inherits(ValueExpression, _Expression);
+
+    function ValueExpression(nodeName, value) {
+        _classCallCheck(this, ValueExpression);
+
+        var _this = _possibleConstructorReturn(this, (ValueExpression.__proto__ || Object.getPrototypeOf(ValueExpression)).call(this));
+
+        _this.value = value;
+        _this.nodeName = nodeName;
+        return _this;
+    }
+
+    _createClass(ValueExpression, [{
+        key: "copy",
+        value: function copy() {
+            return new ValueExpression(this.nodeName, this.value);
+        }
+    }, {
+        key: "isEqualTo",
+        value: function isEqualTo(node) {
+            if (node && this.nodeName === node.nodeName && this.value === node.value) {
+                return true;
+            }
+            return false;
+        }
+    }, {
+        key: "contains",
+        value: function contains(node) {
+            return this.isEqualTo(node);
+        }
+    }]);
+
+    return ValueExpression;
+}(_Expression3.default);
+
+exports.default = ValueExpression;
+//# sourceMappingURL=ValueExpression.js.map
