+++ lib/query/Visitor.js
@@ -14,12 +14,12 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-var _class = function () {
-    function _class() {
-        _classCallCheck(this, _class);
+var Visitor = function () {
+    function Visitor() {
+        _classCallCheck(this, Visitor);
     }
 
-    _createClass(_class, [{
+    _createClass(Visitor, [{
         key: "parse",
         value: function parse(expression) {
             var _this = this;
@@ -41,14 +41,14 @@ var _class = function () {
             var func = this[expression.nodeName];
 
             if (!func) {
-                throw new Error("The builder doesn't support the \"" + expression.nodeName + "\" expression.");
+                throw new Error("The builder doesn't support the \"" + expression.nodeName + '" expression.');
             }
 
             children.forEach(function (child, index) {
-                if (child instanceof BASE.query.Expression) {
+                if (child instanceof _Expression2.default) {
                     var func = _this[child.nodeName];
                     if (!func) {
-                        throw new Error("The builder doesn't support the \"" + child.nodeName + "\" expression.");
+                        throw new Error("The builder doesn't support the \"" + child.nodeName + '" expression.');
                     }
                     children[index] = func.call(_this, child);
                 }
@@ -58,8 +58,8 @@ var _class = function () {
         }
     }]);
 
-    return _class;
+    return Visitor;
 }();
 
-exports.default = _class;
+exports.default = Visitor;
 //# sourceMappingURL=Visitor.js.map
